move models to database package
