+++ depobs/website/config.py
@@ -22,7 +22,7 @@ LOGGING = {
     "loggers": {
         "do": {"handlers": ["console"], "level": "DEBUG"},
         "request.summary": {"handlers": ["console"], "level": "INFO"},
-        "depobs.website.models": {"handlers": ["console"], "level": "INFO"},
+        "depobs.database.models": {"handlers": ["console"], "level": "INFO"},
         "depobs.worker.tasks": {"handlers": ["console"], "level": "INFO"},
         "depobs.worker.scoring": {"handlers": ["console"], "level": "INFO",},
         "depobs.scanner.clients.cratesio": {"handlers": ["console"], "level": "INFO"},
